Introduce interactive mode with -i/--interactive
By default, it would launch an IRB shell on the target process.
A [Rib][] alternative is also included, could be used with:

    rbtrace -p 123 -i rib

This is intended to be pluggable, so that developers could just
provide rbtrace/interactive/ooo so that `-i ooo` would work.
Note that this would use `which` to locate where the command is
and load it. Windows users might need some love to make this work.

Also, I am not an IRB expert and I am not using it, so the patch
here is quite hacky. Please fix it to work with other IRB plugins
if needed.

If you do not like to include the patch for Rib,
I could also provide this from Rib so that the file at:
`lib/rbtrace/interactive/rib.rb` could be removed.

Thanks for considering this! I think this would be quite useful to
figure out what we're really trying to check. The problem is,
we don't want to deploy every time when we want to check something
different, and we're running on Heroku that means we can't really
make changes easily. I haven't tried this on Heroku yet, but I think
this would be useful anyway.

[Rib]: https://github.com/godfat/rib