Find the executable from gem or from PATH
Not sure if this is a good idea or not, maybe we should just
search from PATH since then we don't have to depend on Rubygems
and we need to search from PATH to find `irb` anyway...

Note that `ENV['PATH'].split(':')` probably won't work on Windows.
Oh well, `which` and rbtrace doesn't work on Windows anyway.