Untar msgpack without keeping original permissions
If you're installing rbtrace using the superuser, for example inside
a Docker container, then `tar`'s default behaviour is to add `-p` which
tries to maintain the file ownership/permissions of the original archive

This can cause problems if the folder you've set your Ruby or Bundler
install to use is a network mount to the container or a shared directory
that doesn't understand the permissions requests. It'll lead to an error
unpacking the `msgpack` archive and a failure to install.

By passing in `-o` to the tar command we can skip around the issue and get
a working install in these situations.