Dump heap from a fork
Dumping the heap can lock the VM for a very long time.
If you attempt to do this on a webserver process with a timeout
it's very likely you'll hit it and the process might get killed
before the dump completes.

Worse, if the process was processing some requests or jobs, they might
timeout because of it.

Using fork we can make a snapshot of the heap in a very short time
(relative to the dump itself) and savely dump without disrupting the
original process.

If you are familiar with how Redis does sanopshoting, it's very similar.